Fix expected events behavior

+++ src/globals.h
@@ -64,7 +64,7 @@ extern unsigned g_tw_rng_default; ///< \todo documentation needed
 /** \name Misc Configuration *////@{
 extern unsigned gvt_print_interval;   ///< frequency of progress print outs
 extern unsigned g_tw_stat_interval;   ///< frequency of logging stats (in GVTs)
-extern unsigned g_tw_expected_events; ///< expected number of committed events
+extern long int g_tw_expected_events; ///< expected number of committed events
 ///@}
 
 class EventBuffer;
